# [CS] Solve all conjunctions in source order