Previously we would only do source ordering for ClosureExprs, but other conjunctions need to have their source location taken into account too, in order to make sure we don't try and type-check e.g a TapExpr in a second closure before we type-check the first closure.

Also while here, switch to `std::min_element` instead of sorting, and treat invalid source locations as incomparable.

rdar://113326835