@hamishknight 2023-08-08
@swift-ci please test compiler performance