@hamishknight 2023-08-08
@swift-ci please smoke test compiler performance