@hamishknight 2023-08-09
@swift-ci please smoke test compiler performance