@hamishknight 2023-08-09
@swift-ci please test compiler performance