@hamishknight 2023-08-10
# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-192df930c7533cef)
    - [debug-batch brief](#debug-batch-brief-192df930c7533cef)
    - [debug-batch detailed](#debug-batch-detailed-192df930c7533cef)
- [Release](#release-192df930c7533cef)
    - [release brief](#release-brief-192df930c7533cef)
    - [release detailed](#release-detailed-192df930c7533cef)



<a name="debug-batch-192df930c7533cef">

# Debug-batch

<a name="debug-batch-brief-192df930c7533cef">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 129,558,492,721,662 | 129,502,102,422,498 | -56,390,299,164 | -0.04%
LLVM.NumLLVMBytesOutput | 2,536,493,992 | 2,536,511,888 | 17,896 | 0.0%
</details>

<a name="debug-batch-detailed-192df930c7533cef">

## debug-batch detailed
<details>
<summary>Regressed (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ModuleShadowCacheHit | 68,374 | 69,173 | 799 | 1.17% :no_entry:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.NumSourceLinesPerSecond | 4,791,482 | 4,626,049 | -165,433 | -3.45% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (330)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 4,020,457 | 4,020,298 | -159 | -0.0%
AST.ImportSetCacheMiss | 791,653 | 791,650 | -3 | -0.0%
AST.ImportSetFoldHit | 400,719 | 400,713 | -5 | -0.0%
AST.ImportSetFoldMiss | 390,934 | 390,936 | 2 | 0.0%
AST.ModuleShadowCacheMiss | 5,139 | 5,172 | 33 | 0.64%
AST.ModuleVisibilityCacheHit | 28,935,171 | 28,935,171 | 0 | -0.0%
AST.ModuleVisibilityCacheMiss | 213,067 | 213,067 | 0 | -0.0%
AST.NumASTBytesAllocated | 61,501,686,064 | 61,476,349,477 | -25,336,586 | -0.04%
AST.NumASTScopeExpansions | 8,944,437 | 8,944,407 | -29 | -0.0%
AST.NumASTScopeLookups | 9,962,824 | 9,962,612 | -212 | -0.0%
AST.NumDecls | 350,205 | 350,205 | 0 | 0.0%
AST.NumDependencies | 393,858 | 393,857 | 0 | -0.0%
AST.NumIncrementalDependencies | 25,872 | 25,872 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 320,086 | 320,086 | 0 | 0.0%
AST.NumLocalTypeDecls | 6,622 | 6,622 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,798 | 6,798 | 0 | 0.0%
AST.NumModuleLookupValue | 117,223,106 | 117,220,691 | -2,415 | -0.0%
AST.NumObjCMethods | 95,391 | 95,391 | 0 | 0.0%
AST.NumOperators | 1,028 | 1,028 | 0 | 0.0%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 15,426,627 | 15,426,627 | 0 | 0.0%
AST.NumReferencedTopLevelNames | 1,372,295 | 1,372,295 | 0 | 0.0%
AST.NumSourceBuffers | 397,628 | 397,628 | 0 | 0.0%
AST.NumSourceLines | 12,901,704 | 12,901,704 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 1,819,443 | 1,818,312 | -1,131 | -0.06%
Frontend.MaxMallocUsage | 0 | 0 | 0 | 0.0%
Frontend.NumInstructionsExecuted | 129,558,492,721,662 | 129,502,102,422,498 | -56,390,299,164 | -0.04%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 29,451 | 29,451 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 310,025 | 310,025 | 0 | 0.0%
IRModule.NumIRAliases | 251,232 | 251,232 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 8,631,600 | 8,631,600 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 4,658,370 | 4,658,370 | 0 | 0.0%
IRModule.NumIRGlobals | 4,397,281 | 4,397,281 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 90,090,376 | 90,090,376 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 119,326 | 119,326 | 0 | 0.0%
IRModule.NumIRValueSymbols | 9,135,143 | 9,135,143 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 2,536,493,992 | 2,536,511,888 | 17,896 | 0.0%
Parse.ExportedSourceFileRequest | 4 | 4 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 420,124 | 420,124 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 837,927 | 837,911 | -16 | -0.0%
Parse.ParseAbstractFunctionBodyRequest | 709,463 | 709,463 | 0 | 0.0%
Parse.ParseMembersRequest | 642,705 | 642,689 | -16 | -0.0%
Parse.ParseSourceFileRequest | 372,326 | 372,326 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 372,326 | 372,326 | 0 | 0.0%
SILGen.ASTLoweringRequest | 30,635 | 30,635 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 3,675,500 | 3,675,500 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 143,341 | 143,341 | 0 | 0.0%
SILModule.NumSILGenVtables | 40,663 | 40,663 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 213,927 | 213,927 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 4,188,934 | 4,188,934 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 145,406 | 145,406 | 0 | 0.0%
SILModule.NumSILOptVtables | 41,032 | 41,032 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 258,630 | 258,630 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 150,837 | 150,837 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 319,740 | 319,739 | -1 | -0.0%
Sema.AbstractGenericSignatureRequest | 151,272 | 151,186 | -85 | -0.06%
Sema.AccessLevelRequest | 19,513,771 | 19,475,806 | -37,965 | -0.19%
Sema.ActorIsolationRequest | 7,273,695 | 7,273,699 | 4 | 0.0%
Sema.AllMembersRequest | 179,370 | 179,370 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 4,568,963 | 4,568,684 | -279 | -0.01%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 50,596 | 50,596 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 33,130 | 33,130 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 23,561,595 | 23,536,158 | -25,436 | -0.11%
Sema.AttachedResultBuilderRequest | 1,976,191 | 1,976,191 | 0 | 0.0%
Sema.BodyInitKindRequest | 99,761 | 99,761 | 0 | 0.0%
Sema.BreakTargetRequest | 6,936 | 6,936 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 496,997 | 496,997 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 59,425 | 59,425 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,438 | 10,438 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 59,425 | 59,425 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,438 | 10,438 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 2,442,886 | 2,442,886 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 20,091 | 20,091 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 20,091 | 20,091 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 154,430 | 154,409 | -21 | -0.01%
Sema.ClosureEffectsRequest | 156,695 | 156,695 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 60,162 | 60,162 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 15,689,588 | 15,653,894 | -35,694 | -0.23%
Sema.CompareDeclSpecializationRequest | 894,021 | 893,705 | -316 | -0.04%
Sema.CompilerPluginLoadRequest | 4 | 4 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 907,385 | 907,231 | -154 | -0.02%
Sema.ConformanceHasEffectRequest | 63 | 63 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 2,293 | 2,293 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 23,346 | 23,346 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 3,690 | 3,690 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 71,740 | 71,683 | -57 | -0.08%
Sema.DefaultAndMaxAccessLevelRequest | 126,031 | 126,031 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 111,428 | 111,428 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 884 | 884 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 22,160 | 22,160 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 8,661 | 8,661 | 0 | 0.0%
Sema.DefaultTypeRequest | 678,156 | 678,156 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 86,228,503 | 86,215,902 | -12,600 | -0.01%
Sema.DirectOperatorLookupRequest | 2,582,482 | 2,582,492 | 9 | 0.0%
Sema.DirectPrecedenceGroupLookupRequest | 958,440 | 958,440 | 0 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 7,755,025 | 7,755,029 | 4 | 0.0%
Sema.EnumElementExprPatternRequest | 19 | 19 | 0 | 0.0%
Sema.EnumRawTypeRequest | 43,017 | 43,017 | 0 | 0.0%
Sema.EnumRawValuesRequest | 25,233 | 25,233 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 34,301 | 34,298 | -2 | -0.01%
Sema.ExpandAccessorMacros | 1,103,822 | 1,103,822 | 0 | 0.0%
Sema.ExpandChildTypeRefinementContextsRequest | 729,018 | 729,018 | 0 | 0.0%
Sema.ExpandExtensionMacros | 1,075,047 | 1,074,291 | -756 | -0.07%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 70,751,317 | 70,663,672 | -87,645 | -0.12%
Sema.ExpandPeerMacroRequest | 50,739,775 | 50,646,519 | -93,256 | -0.18%
Sema.ExpandSynthesizedMemberMacroRequest | 317,629 | 317,628 | -1 | -0.0%
Sema.ExprPatternMatchRequest | 99,191 | 99,191 | 0 | 0.0%
Sema.ExtendedNominalRequest | 735,672 | 735,672 | 0 | 0.0%
Sema.ExtendedTypeRequest | 226,872 | 226,871 | -1 | -0.0%
Sema.ExternalMacroDefinitionRequest | 4 | 4 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,906,558 | 2,906,558 | 0 | 0.0%
Sema.FunctionOperatorRequest | 20,626 | 20,626 | 0 | 0.0%
Sema.GenericParamListRequest | 7,486,945 | 7,484,854 | -2,091 | -0.03%
Sema.GenericSignatureRequest | 3,697,971 | 3,697,492 | -479 | -0.01%
Sema.GetDestructorRequest | 47,356 | 47,356 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 267,566 | 267,530 | -36 | -0.01%
Sema.GetRuntimeDiscoverableAttributes | 2,568,591 | 2,568,591 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 138 | 138 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 9,241,326 | 9,241,423 | 96 | 0.0%
Sema.GlobalActorInstanceRequest | 1,127 | 1,127 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 18,995 | 18,995 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 26,138 | 26,138 | 0 | 0.0%
Sema.HasDefaultInitRequest | 110,855 | 110,855 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 32,472 | 32,472 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 589,561 | 589,475 | -86 | -0.01%
Sema.HasImportsMatchingFlagRequest | 738,291 | 738,291 | 0 | 0.0%
Sema.HasInitAccessorRequest | 5,712,250 | 5,704,615 | -7,634 | -0.13%
Sema.HasIsolatedSelfRequest | 16,433,774 | 16,423,084 | -10,690 | -0.07%
Sema.HasMemberwiseInitRequest | 55,871 | 55,871 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 33,552 | 33,552 | 0 | 0.0%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 17,000 | 17,000 | 0 | 0.0%
Sema.HasStorageRequest | 2,421,260 | 2,421,260 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 110,865 | 110,865 | 0 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 249,111 | 249,106 | -5 | -0.0%
Sema.InheritedDeclsReferencedRequest | 7,959,280 | 7,952,839 | -6,441 | -0.08%
Sema.InheritedProtocolsRequest | 854,085 | 853,972 | -113 | -0.01%
Sema.InheritedTypeRequest | 490,805 | 490,752 | -53 | -0.01%
Sema.InheritsSuperclassInitializersRequest | 42,507 | 42,507 | 0 | 0.0%
Sema.InitAccessorPropertiesRequest | 51,595 | 51,595 | 0 | 0.0%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 208,613 | 208,613 | 0 | 0.0%
Sema.InterfaceTypeRequest | 24,969,279 | 24,942,574 | -26,704 | -0.11%
Sema.IsABICompatibleOverrideRequest | 342,697 | 342,697 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 614,257 | 614,257 | 0 | 0.0%
Sema.IsActorRequest | 2,518,088 | 2,517,344 | -743 | -0.03%
Sema.IsCallAsFunctionNominalRequest | 7,010 | 7,010 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 72,592 | 72,592 | 0 | 0.0%
Sema.IsDefaultActorRequest | 64,423 | 64,423 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,811,633 | 2,810,788 | -845 | -0.03%
Sema.IsDynamicRequest | 3,458,005 | 3,458,005 | 0 | 0.0%
Sema.IsFinalRequest | 5,212,484 | 5,208,060 | -4,424 | -0.08%
Sema.IsGetterMutatingRequest | 870,727 | 870,727 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 4,241,968 | 4,241,642 | -326 | -0.01%
Sema.IsMoveOnlyRequest | 798,260 | 798,219 | -41 | -0.01%
Sema.IsNonUserModuleRequest | 103,411 | 103,426 | 14 | 0.01%
Sema.IsObjCRequest | 3,210,192 | 3,209,913 | -279 | -0.01%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 781,870 | 781,870 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 396,658 | 396,658 | 0 | 0.0%
Sema.IsStaticRequest | 1,732,872 | 1,732,825 | -46 | -0.0%
Sema.LazyStoragePropertyRequest | 2,585 | 2,585 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 586,811 | 586,811 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 3,232,963 | 3,227,905 | -5,057 | -0.16%
Sema.LookupConformanceInModuleRequest | 130,256,850 | 130,213,759 | -43,091 | -0.03%
Sema.LookupInModuleRequest | 10,040,223 | 10,039,891 | -332 | -0.0%
Sema.LookupInfixOperatorRequest | 83,413 | 83,413 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 34,012 | 34,012 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 454 | 454 | 0 | 0.0%
Sema.MacroDefinitionRequest | 8 | 8 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 3,192 | 3,192 | 0 | 0.0%
Sema.MemberwiseInitPropertiesRequest | 51,597 | 51,597 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 10,438 | 10,438 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 22 | 22 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 3,778,255 | 3,778,116 | -139 | -0.0%
Sema.NamedLazyMemberLoadSuccessCount | 66,273,580 | 66,274,012 | 432 | 0.0%
Sema.NamingPatternRequest | 254,300 | 254,299 | -1 | -0.0%
Sema.NeedsNewVTableEntryRequest | 1,113,232 | 1,113,232 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 496,497 | 496,497 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 598,092 | 598,092 | 0 | 0.0%
Sema.NumConformancePathsRecorded | 1,014,190 | 1,014,198 | 8 | 0.0%
Sema.NumConformancesDeserialized | 6,299,672 | 6,294,709 | -4,963 | -0.08%
Sema.NumConstraintScopes | 79,294,610 | 79,293,320 | -1,290 | -0.0%
Sema.NumConstraintsConsideredForEdgeContraction | 2,814,275 | 2,814,275 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 98,657,255 | 98,571,398 | -85,856 | -0.09%
Sema.NumDeclsTypechecked | 3,471,318 | 3,471,318 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 14,761,459 | 14,759,491 | -1,968 | -0.01%
Sema.NumLazyRequirementSignatures | 1,119,685 | 1,119,655 | -30 | -0.0%
Sema.NumLazyRequirementSignaturesLoaded | 788,067 | 788,032 | -35 | -0.0%
Sema.NumLeafScopes | 61,099,108 | 61,098,008 | -1,100 | -0.0%
Sema.NumRequirementMachineCompletionSteps | 27,569,331 | 27,559,057 | -10,274 | -0.04%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,364,176 | 1,366,167 | 1,991 | 0.15%
Sema.NumRequirementMachines | 2,348,783 | 2,347,860 | -923 | -0.04%
Sema.NumTypesDeserialized | 27,334,006 | 27,325,101 | -8,904 | -0.03%
Sema.NumUnloadedLazyIterableDeclContexts | 11,217,117 | 11,220,827 | 3,709 | 0.03%
Sema.ObjCInterfaceAndImplementationRequest | 986,392 | 986,389 | -2 | -0.0%
Sema.OpaqueReadOwnershipRequest | 531,224 | 531,224 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 690 | 690 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,027 | 1,027 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 6,849,788 | 6,846,746 | -3,042 | -0.04%
Sema.ParamSpecifierRequest | 2,354,073 | 2,354,072 | -1 | -0.0%
Sema.PatternBindingEntryRequest | 859,497 | 859,496 | -1 | -0.0%
Sema.PatternTypeRequest | 1,016,898 | 1,016,897 | -1 | -0.0%
Sema.PolymorphicEffectKindRequest | 1,154,364 | 1,154,364 | 0 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 75 | 75 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 7,479,401 | 7,479,200 | -201 | -0.0%
Sema.PreCheckFunctionBodyRequest | 465,846 | 465,846 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 1,169 | 1,169 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 633,096 | 633,096 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 20,011 | 19,988 | -23 | -0.11%
Sema.PrimarySourceFilesRequest | 10,438 | 10,438 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,582,256 | 2,582,256 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 33,130 | 33,130 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 754,874 | 754,874 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 2,839 | 2,839 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 863,641 | 863,641 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 651 | 651 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 677,061 | 676,976 | -85 | -0.01%
Sema.ProtocolRequirementsRequest | 83,323 | 83,322 | -1 | -0.0%
Sema.ProtocolRequiresClassRequest | 48,897 | 48,830 | -67 | -0.14%
Sema.ProvideDefaultImplForRequest | 15,689,588 | 15,653,894 | -35,694 | -0.23%
Sema.QualifiedLookupRequest | 10,489,283 | 10,487,194 | -2,088 | -0.02%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 1,268,502 | 1,268,502 | 0 | 0.0%
Sema.RenamedDeclRequest | 5,542 | 5,542 | 0 | 0.0%
Sema.RequirementRequest | 197,549 | 197,548 | -1 | -0.0%
Sema.RequirementSignatureRequest | 848,543 | 848,454 | -89 | -0.01%
Sema.RequiresOpaqueAccessorsRequest | 2,199,679 | 2,199,679 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 534,023 | 534,023 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveExtensionMacroConformances | 8 | 8 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 1,041,145 | 1,040,889 | -256 | -0.02%
Sema.ResolveMacroRequest | 27,328 | 27,340 | 12 | 0.04%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveRawLayoutLikeTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 5,974,385 | 5,974,362 | -23 | -0.0%
…